feat: Add post service

+++ src/main/java/me/mocha/minisns/model/dto/PostDTO.java
@@ -21,7 +21,7 @@ public class PostDTO {
     private String content;
 
     @Getter
-    private int views;
+    private long views;
 
     @Getter
     private String username;
